requirements.txt
attrs
git+https://github.com/TomGoBravo/flask-admin@nocustomtiles
geojson>=2.4.1
SQLAlchemy-Continuum
Flask
Flask-Admin
Flask-SQLAlchemy
Flask-Markdown
Flask-Migrate
shapely
flask-pagedown
git+https://github.com/geoalchemy/geoalchemy2@master
polyline
psycopg2
uritemplate
blinker
flask-login
flask_dance
sqlalchemy
pytest
